+++ src/main/java/com/gofobao/framework/tender/vo/response/VoFindAutoTender.java
@@ -2,6 +2,7 @@ package com.gofobao.framework.tender.vo.response;
 
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 /**
@@ -31,9 +32,9 @@ public class VoFindAutoTender {
     private Date autoAt;
     private Date createdAt;
     private Date updatedAt;
-    private Integer useMoney;
-    private Integer noUseMoney;
-    private Integer virtualMoney;
-    private Integer collection;
-    private Integer payment;
+    private BigInteger useMoney;
+    private BigInteger  noUseMoney;
+    private BigInteger  virtualMoney;
+    private BigInteger  collection;
+    private BigInteger  payment;
 }
